trainer/requirements.txt
ultralytics==8.3.0
opencv-python-headless==4.9.0.80
google-cloud-storage==2.18.0
PyYAML==6.0.2